Skyler Sebold - Added DemoSwitchSubsystem
Created the DemoSwitchSubsystem, its corresponding constants, and demospeed constants for Shooter and Intake (Unknown if needed for intake, and for Drive sub due to the workings of FRC's Tank Drive class, instead of different speed constants, DemoSwitchSub's getter will instead determine whether to pass the controller joystick Y positions directly, or to add dampeners to them first [i.e. LeftStickY * 0.5]), The implementations and logic for the DemoSwitch is still yet to be implemented into the corresponding affected subsystems using static accessors (DemoSwitchSub::GetDemoSwitch)
furthermore for the purpose of me not forgetting what comes next, the function in the intake subsystem for loading to the photogate is yet to be created and must still be made.